+++ src/query/storages/fuse/Cargo.toml
@@ -44,7 +44,6 @@ backoff = { version = "0.4.0", features = ["futures", "tokio"] }
 chrono = { workspace = true }
 futures = "0.3.24"
 futures-util = "0.3.24"
-itertools = "0.10.5"
 metrics = "0.20.1"
 opendal = { workspace = true }
 serde = { workspace = true }
