+++ src/query/storages/fuse/src/operations/mutation/compact/mod.rs
@@ -14,21 +14,16 @@
 
 mod block_compact_mutator;
 mod compact_meta;
-mod compact_mutator;
 mod compact_part;
 mod compact_source;
-mod compact_source2;
-mod compact_transform;
-mod merge_segments_transform;
 mod segment_compact_mutator;
 mod transform_compact_aggregator;
 
 pub use block_compact_mutator::BlockCompactMutator;
-pub use compact_meta::CompactSinkMeta;
-pub use compact_part::CompactPartInfo2;
+pub use compact_meta::CompactSourceMeta;
+pub use compact_part::CompactPartInfo;
 pub use compact_source::CompactSource;
-pub use compact_transform::CompactTransform;
-pub use merge_segments_transform::MergeSegmentsTransform;
 pub use segment_compact_mutator::SegmentCompactMutator;
 pub use segment_compact_mutator::SegmentCompactionState;
 pub use segment_compact_mutator::SegmentCompactor;
+pub use transform_compact_aggregator::CompactAggregator;
