+++ src/query/storages/fuse/src/operations/mutation/mod.rs
@@ -26,10 +26,9 @@ mod serialize_data_transform;
 pub use abort_operation::AbortOperation;
 pub use base_mutator::BaseMutator;
 pub use compact::BlockCompactMutator;
-pub use compact::CompactPartInfo2;
+pub use compact::CompactAggregator;
+pub use compact::CompactPartInfo;
 pub use compact::CompactSource;
-pub use compact::CompactTransform;
-pub use compact::MergeSegmentsTransform;
 pub use compact::SegmentCompactMutator;
 pub use compact::SegmentCompactionState;
 pub use compact::SegmentCompactor;
